Deck selector implementation and SQL organization

+++ import_csv_file.py
@@ -4,55 +4,45 @@ from datetime import datetime
 
 date_now = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
 
-conn = sqlite3.connect("decks.db")
+conn = sqlite3.connect("Kotoba.db")
 cursor = conn.cursor()
 
-cursor.execute("DROP TABLE IF EXISTS Decks;")
-cursor.execute("DROP TABLE IF EXISTS Cards;")
-cursor.execute("DROP TABLE IF EXISTS CardReviews;")
+if input() == "y":
+    cursor.execute("DROP TABLE IF EXISTS Decks;")
+    cursor.execute("DROP TABLE IF EXISTS Cards;")
 
-cursor.execute(
-    """
-    CREATE TABLE Decks (
-        deck_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        name TEXT NOT NULL,
-        description TEXT,
-        created_at DATETIME NOT NULL
-    );
-    """
-)
+    cursor.execute(
+        """
+        CREATE TABLE Decks (
+            deck_id INTEGER PRIMARY KEY AUTOINCREMENT,
+            name TEXT NOT NULL,
+            description TEXT,
+            created_at DATETIME NOT NULL
+        );
+        """
+    )
 
-cursor.execute(
-    """
-    CREATE TABLE Cards (
-        card_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        deck_id INTEGER NOT NULL,
-        front TEXT NOT NULL,
-        reading TEXT,
-        back TEXT NOT NULL,
-        image_url TEXT,
-        created_at DATETIME NOT NULL,
-        FOREIGN KEY (deck_id) REFERENCES Decks(deck_id) ON DELETE CASCADE
-    );
-    """
-)
-
-cursor.execute(
-    """
-    CREATE TABLE CardReviews (
-        review_id INTEGER PRIMARY KEY AUTOINCREMENT,
-        card_id INTEGER NOT NULL,
-        next_review_date DATETIME DEFAULT CURRENT_TIMESTAMP,
-        FOREIGN KEY (card_id) REFERENCES Cards(card_id) ON DELETE CASCADE
-    );
-    """
-)
+    cursor.execute(
+        """
+        CREATE TABLE Cards (
+            card_id INTEGER PRIMARY KEY AUTOINCREMENT,
+            deck_id INTEGER NOT NULL,
+            front TEXT NOT NULL,
+            reading TEXT,
+            back TEXT NOT NULL,
+            image_url TEXT,
+            next_review_date DATETIME DEFAULT CURRENT_TIMESTAMP,
+            created_at DATETIME NOT NULL,
+            FOREIGN KEY (deck_id) REFERENCES Decks(deck_id) ON DELETE CASCADE
+        );
+        """
+    )
 
 csv_file_path = "Dareka no Manazashi.csv"
 
 cursor.execute(
     "INSERT INTO Decks (name, description, created_at) VALUES (?, ?, ?)",
-    (csv_file_path, "mami", date_now),
+    (csv_file_path.removesuffix(".csv"), "A japanese deck", date_now),
 )
 
 deck_id = cursor.lastrowid
@@ -68,13 +58,6 @@ with open(csv_file_path, newline="", encoding="utf-8") as csvfile:
         cursor.execute(
             "INSERT INTO Cards (deck_id, front, reading, back, image_url, created_at) VALUES (?, ?, ?, ?, ?,?);",
             (deck_id, front, reading, back, img_url, date_now),
-        )
-
-        card_id = cursor.lastrowid
-
-        cursor.execute(
-            "INSERT INTO CardReviews (card_id, next_review_date) VALUES (?, ?);",
-            (card_id, date_now),
         )
 
 conn.commit()
